Impl Iterator for Lexer and remove Range trait(s)
The Lexer now implements the Iterator trait so that spans can be tracked
and displayed correctly.

The Range<usize> trait was removed from the API as Spans can instead be
constructed from two usize values. This allows Span to implement Copy
without having to worry about anything related to the
RangeBounds-related traits (some of which are not Copy).