Added whitespace handling and extended the rules
The lexer uses a set of rules to determine how to handle particular
characters, and/or sequences of characters. This is what allows it to
convert symbols into different kinds of tokens.

The rules will be extended further to include more general character
sequences, such as identifiers, integers, comments, etc.

Fixed a few issues RE: missing or mismatched token mappings.